+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## [Unreleased]
 
+### Fixed
+
+- random_seed is lost when changing page [#504](https://github.com/spraakbanken/korp-frontend/issues/504)
+
 ## [9.14.2] – 2026-02-10
 
 ### Fixed